02.05/2026

+++ sdk/include/source2sdk/server/HandshakeState_t.hpp
@@ -14,9 +14,9 @@ namespace source2sdk
     namespace server
     {
         // Enumerator count: 7
-        // Alignment: 4
-        // Size: 0x4
-        enum class HandshakeState_t : std::uint32_t
+        // Alignment: 1
+        // Size: 0x1
+        enum class HandshakeState_t : std::uint8_t
         {
             eInactive = 0x0,
             eWaitingToBecomeInactive = 0x1,
